+++ server/controllers/apiTradeExchange.js
@@ -72,7 +72,11 @@ exports.readAllTradeExchange = async (req, res) => {
 };
 
   exports.deleteTradeExchange = async (req, res) => {
-  
+    
+    if ('ADMIN' != req.user.role) {
+      res.status(401).json({error: "Do not have permission"})
+      return;
+    }
       try {
         const tradeExchange = await TradeExchange.findByIdAndDelete({_id: req.params.id});
